+++ lib/international_phone_text_field.dart
@@ -1,3 +1,4 @@
 library;
 
 export 'src/international_phone_text_field.dart';
+export 'src/international_phone_text_field_theme.dart';