+++ lib/src/utils/wrong_flag_container.dart
@@ -1,9 +1,13 @@
 import 'package:flutter/material.dart';
+import 'package:international_phone_text_field/src/international_phone_text_field_theme.dart';
 
 class WrongFlagContainer extends StatelessWidget {
   const WrongFlagContainer({
     super.key,
+    this.theme = const InternationalPhoneTextFieldTheme(),
   });
+
+  final InternationalPhoneTextFieldTheme theme;
 
   @override
   Widget build(BuildContext context) {
@@ -11,11 +15,11 @@ class WrongFlagContainer extends StatelessWidget {
       height: 18,
       width: 34,
       decoration: BoxDecoration(
-        color: Colors.grey.shade300,
+        color: theme.unknownFlagColor,
         borderRadius: BorderRadius.circular(3),
       ),
       alignment: Alignment.center,
-      child: Text("?"),
+      child: Text("?", style: theme.unknownFlagTextStyle),
     );
   }
 }